Combine tests into a single suite. Assert against the generated SQL and run the query. Don't check the rows returned by the query.
