+++ .github/workflows/ci.yml
@@ -126,15 +126,8 @@ jobs:
           ruby-version: ${{ matrix.ruby-version }}
           bundler-cache: true
 
-      - name: Run unit tests
-        env:
-          COVERAGE_COMMAND_NAME: unit
+      - name: Run tests
         run: bundle exec rake test
-
-      - name: Run integration tests
-        env:
-          COVERAGE_COMMAND_NAME: integration
-        run: bundle exec rake test:integration
 
       - name: Upload to Codecov
         uses: codecov/codecov-action@v7
