+++ spec/index_generator_spec.rb
