+++ spec/key_field_spec.rb
@@ -17,10 +17,8 @@ class RuntimeKeyDocIndex < ParadeDB::Index
   }
 end
 
-RSpec.describe "KeyFieldRuntimeIntegrationTest" do
+RSpec.describe "KeyFieldRuntime" do
   before do
-    skip "Integration tests require PostgreSQL" unless postgresql?
-
     ActiveRecord::Schema.define do
       suppress_messages do
         create_table :runtime_key_docs, force: true do |t|
@@ -35,36 +33,33 @@ RSpec.describe "KeyFieldRuntimeIntegrationTest" do
     conn.execute("CREATE EXTENSION IF NOT EXISTS pg_search CASCADE;")
     conn.remove_paradedb_index(:runtime_key_docs, name: :runtime_key_docs_search_idx, if_exists: true)
     conn.create_paradedb_index(RuntimeKeyDocIndex)
-
-    RuntimeKeyDoc.connection.execute("TRUNCATE TABLE runtime_key_docs RESTART IDENTITY;")
-    RuntimeKeyDoc.create!(external_id: 101, body: "wireless bluetooth earbuds", tag: "audio")
-    RuntimeKeyDoc.create!(external_id: 102, body: "wireless noise cancelling headphones", tag: "audio")
-    RuntimeKeyDoc.create!(external_id: 103, body: "running shoes lightweight", tag: "footwear")
   end
 
   after do
-    next unless postgresql?
-
     conn = ActiveRecord::Base.connection
     conn.remove_paradedb_index(:runtime_key_docs, name: :runtime_key_docs_search_idx, if_exists: true) rescue nil
     conn.drop_table(:runtime_key_docs, if_exists: true) rescue nil
   end
 
   it "with_score works with non-primary-key key_field" do
-    rows = RuntimeKeyDoc.search(:body)
-                        .match_all("wireless")
-                        .with_score
-                        .order(search_score: :desc)
-                        .limit(3)
-                        .to_a
+    sql = RuntimeKeyDoc.search(:body).match_all("wireless").with_score.order(search_score: :desc).limit(3).to_sql
 
-    refute_empty rows
-    rows.each { |row| assert_operator row.search_score.to_f, :>=, 0.0 }
+    assert_query_sql <<~SQL, sql
+      SELECT runtime_key_docs.*, pdb.score("runtime_key_docs"."external_id") AS search_score FROM runtime_key_docs
+      WHERE ("runtime_key_docs"."body" &&& 'wireless')
+      ORDER BY search_score DESC
+      LIMIT 3
+    SQL
   end
 
   it "more_like_this uses DSL key_field value" do
-    ids = RuntimeKeyDoc.more_like_this(101, fields: [:body]).order(:external_id).pluck(:external_id)
-    assert_includes ids, 102
+    sql = RuntimeKeyDoc.more_like_this(101, fields: [:body]).order(:external_id).to_sql
+
+    assert_query_sql <<~SQL, sql
+      SELECT runtime_key_docs.* FROM runtime_key_docs
+      WHERE ("runtime_key_docs"."external_id" @@@ pdb.more_like_this(101, ARRAY['body']))
+      ORDER BY "runtime_key_docs"."external_id" ASC
+    SQL
   end
 
   it "with_facets without ParadeDB predicates works with non-primary-key key_field" do
@@ -74,17 +69,11 @@ RSpec.describe "KeyFieldRuntimeIntegrationTest" do
                             .order(:external_id)
                             .limit(10)
 
-    rows = relation.to_a
-    refute_empty rows
-
-    facets = relation.facets
-    assert_kind_of Hash, facets
-    assert_includes facets, "agg"
-  end
-
-  private
-
-  def postgresql?
-    ActiveRecord::Base.connection.adapter_name.downcase.include?("postgres")
+    assert_query_sql <<~SQL, relation.to_sql
+      SELECT runtime_key_docs.*, pdb.agg('{"value_count":{"field":"external_id"}}') OVER () AS _agg_facet FROM runtime_key_docs
+      WHERE "runtime_key_docs"."tag" = 'audio' AND ("runtime_key_docs"."external_id" @@@ pdb.all())
+      ORDER BY "runtime_key_docs"."external_id" ASC
+      LIMIT 10
+    SQL
   end
 end
