+++ spec/proximity_spec.rb
